Update stale branch 187 with master

See merge request carboncopies/BrainGenix-NES!203

+++ Source/Core/VSDA/EM/NeuroglancerConversionPool/NeuroglancerConverter.cpp
@@ -136,7 +136,7 @@ bool ExecuteConversionOperation(BG::Common::Logger::LoggingSystem* _Logger, Simu
     //  - Create the resolution sizes
     int ResX_nm = Params->VoxelResolution_um * 1000;
     int ResY_nm = Params->VoxelResolution_um * 1000;
-    int ResZ_nm = int(Params->SliceThickness_um / Params->VoxelResolution_um) * 1000 * Params->VoxelResolution_um;
+    int ResZ_nm = (Params->SliceThickness_um / Params->VoxelResolution_um) * 1000 * Params->VoxelResolution_um;
     std::vector<int> Resolution{ResX_nm, ResY_nm, ResZ_nm};
     Scales["resolution"] = Resolution;
